LayoutTests/fast/css/getComputedStyle/getComputedStyle-background-size-expected.txt

Test calling getPropertyValue on computed styles for -webkit-backround-size properties.

getPropertyValue(-webkit-background-size): 10px 20px
getPropertyCSSValue(-webkit-background-size): [object CSSValueList]
getPropertyValue(-webkit-background-size): 10px 10px
getPropertyCSSValue(-webkit-background-size): [object CSSValueList]

Test getting background size of 0.
getPropertyValue(-webkit-background-size): 0px 0px
getPropertyCSSValue(-webkit-background-size): [object CSSValueList]
getPropertyValue(-webkit-background-size): 0px 0px
getPropertyCSSValue(-webkit-background-size): [object CSSValueList]

Test getting initial value of the background size.
getPropertyValue(-webkit-background-size): auto
getPropertyCSSValue(-webkit-background-size): [object CSSPrimitiveValue]
